% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rocForest.R
\name{predict.rocForest}
\alias{predict.rocForest}
\title{Predicting based on a \code{rocForest} model.}
\usage{
\method{predict}{rocForest}(object, newdata, type = c("survival", "hazard",
  "cumHaz"), ...)
}
\arguments{
\item{object}{is an \code{rocForest} object.}

\item{newdata}{is an optional data frame in which to look for variables with which to predict.
If omitted, the fitted predictors are used.
If the covariate observation time is not supplied, covariates will be treated as at baseline.}

\item{type}{is an optional character string specifying whether to predict the survival probability or the hazard rate.}

\item{...}{for future developments.}
}
\description{
The function gives predicted values.
}
\seealso{
\code{\link{predict.rocTree}}
}
