% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rocForest.R
\name{print.rocForest}
\alias{print.rocForest}
\title{Printing an \code{rocForest} object}
\usage{
\method{print}{rocForest}(x, tree = 1L, digits = 5, dt = TRUE, ...)
}
\arguments{
\item{x}{an \code{rocForest} object.}

\item{tree}{an integer specifying the \eqn{n^{th}} tree in the forest to print.
The function prints the first tree in the forest by default.}

\item{digits}{the number of digits of numbers to print.}

\item{dt}{an optional logical vector. If TRUE, tree structure based on \strong{\code{data.tree} structure} is printed.}

\item{...}{for future development.}
}
\description{
Prints a tree from an \code{rocForest} object.
}
