% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocTree_pkg.R
\docType{package}
\name{rocTree-package}
\alias{rocTree-package}
\alias{_PACKAGE}
\title{rocTree:Receiver Operating Characteristic (ROC)-Guided Classification and Survival Tree}
\description{
\code{rocTree} uses a ROC-guided classification algorithm to grow and prune survival trees.
Hazards at terminal notes are also computed.
}
\section{Introduction}{

Existing classification, regression, and survival trees are typically implemented via a
greedy algorithm that maximizes the within node homogeneity or between node heterogeneity.
Sun and Wang (2018+) re-define the ROC curves for tree classifiers using the idea of
randomized test and show that the target function yields the highest ROC curve.
Then a criterion based on ROC is used in the tree-building algorithm.
}

\section{Methods}{

The package contains functions to construct ROC-guided survival trees (\code{\link{rocTree}}).
}

\examples{
library(survival)
data(simudat)
system.time(fit <- rocTree(Surv(Time, Status) ~ X1 + X2, id = ID, data = simudat,
control = list(CV = TRUE, nflds = 10)))
fit

}
\seealso{
\code{\link{rocTree}}

For more details, see the \code{rocTree} vignette by running: \code{vignette("rocTree")}
}
\author{
\strong{Maintainer}: Sy Han Chiou \email{schiou@utdallas}

Authors:
\itemize{
  \item Yifei Sun \email{ys3072@cumc.columbia.edu}
  \item Mei-Cheng Wang \email{mcwang@jhu.edu}
}

}
