% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocTree.R
\name{rocTree.Xlist}
\alias{rocTree.Xlist}
\title{Prepare xlist in \code{rocTree} and \code{rocForest}}
\usage{
rocTree.Xlist(x, disc, y, id)
}
\arguments{
\item{x}{is a covariate vector.}

\item{disc}{a binary value indicating whether \code{x} is disctete;
\code{disc = 1} if discrete.}

\item{y}{is the ordered event time observed in the data.}

\item{id}{subject's id}
}
\description{
An internal function used to prepare \code{x.list}.
This is exported for now to provide transformed covariate path
for fitting rpart and ctree.
}
\details{
Should this be make public?
}
