% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simDat}
\alias{simDat}
\title{Simulated dataset for demonstration}
\format{
A data frame with 5050 rows and 5 variables.
}
\description{
A simulated data frame with variables
\describe{
  \item{id}{subjects identification}
  \item{Time}{observation times}
  \item{death}{event/death indicator; 1 = an event (death) is recorded}
  \item{z1}{baseline covariate generated from a standard uniform distribution}
  \item{z2}{baseline covariate generated from a standard uniform distribution (independent from \code{z1}}
}
}
\details{
The sample dataset is generated by \code{set.seed(1); simu(100, 0, 1.3)}.
See \code{\link{simu}} for details of the \code{simu} function.
}
