% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudat.R
\docType{data}
\name{simudat}
\alias{simudat}
\title{Sample Data: A simulated data.}
\format{A data.frame with 4046 rows and 6 columns.
\describe{
  \item{ID}{patient's ID}
  \item{Time}{patient's observed times}
  \item{Status}{censoring indicator, 0 for censored, 1 for event.}
  \item{X1}{time independent variable}
  \item{X2}{time dependent variable}
  \item{X3}{another time independent variable}
}}
\usage{
data(simudat)
}
\description{
A simulated data with both time independent and dependent covariate.
}
