% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTree.R
\name{splitRP}
\alias{splitRP}
\title{This is an internal function, called by \code{grow.RP}. Inputs are based on an ordered Y.}
\usage{
splitRP(X, Y, E, fmat, Smat, treeMat, ndInd, const, fTree, STree, parm,
  randP = NULL)
}
\description{
This is an internal function, called by \code{grow.RP}. Inputs are based on an ordered Y.
}
