% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{trueHaz}
\alias{trueHaz}
\title{Function to generate the true hazard used in the simulation.}
\usage{
trueHaz(tt, scenario)
}
\arguments{
\item{tt}{is a numerical vector, specifying where to evaluate at.}

\item{scenario}{can be numeric or character string.
This indicates the simulation scenario noted in the manuscript. See \code{simu} for details.}
}
\description{
This function is used to generate the true cumulative hazard function used in the simulation.
}
