% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{trueSurv}
\alias{trueSurv}
\title{Function to generate the true survival used in the simulation.}
\usage{
trueSurv(tt, scenario)
}
\arguments{
\item{tt}{is a numerical vector, specifying where to evaluate at.}

\item{scenario}{can be numeric or character string.
This indicates the simulation scenario noted in the manuscript. See \code{simu} for details.}
}
\description{
This function is used to generate the true survival function used in the simulation.
}
