% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rocForest.R
\name{plot.rocForest}
\alias{plot.rocForest}
\title{Plotting an \code{rocForest} object}
\usage{
\method{plot}{rocForest}(x, tree = 1L, output = c("graph", "visNetwork"),
  digits = 4, control = list(), ...)
}
\arguments{
\item{x}{an \code{rocForest} object.}

\item{tree}{an integer specifying the \eqn{n^{th}} tree in the forest to print.}

\item{output}{a string specifying the output type; graph (the default) renders the graph using the \code{grViz} function, and \code{visNetwork} renders the graph using the visnetwork function.}

\item{digits}{the number of digits to print.}

\item{control}{a list of control parameters. See "details" for important features of control parameters.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Plots a tree from an \code{rocForest} object.
}
