% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rocTree.R
\name{plot.rocTree}
\alias{plot.rocTree}
\title{Plotting an \code{rocTree} object}
\usage{
\method{plot}{rocTree}(x, output = c("graph", "visNetwork"),
  digits = 4, control = list(), ...)
}
\arguments{
\item{x}{an object of class "\code{rocTree}", usually returned by the rocTree function.}

\item{output}{a string specifying the output type; graph (the default) renders the graph using the \code{grViz} function, and \code{visNetwork} renders the graph using the visnetwork function.}

\item{digits}{the number of digits to print.}

\item{control}{a list of control parameters. See "details" for important features of control parameters.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Plots an \code{rocTree} object. The function uses a \code{dgr_graph} object, render the graph in the RStudio Viewer.
}
\details{
The argument \code{control} consists of some options to be passed down to \code{render_graph} of the \code{DiagrammeR} package.
The argument defaults to a list with the following values:
\describe{
  \item{rankdir}{a character vector specifying the drawing direction.
The default value is \code{TB} (top-to-bottom).
Other options include \code{BT} (bottom-to-top), \code{LR} (left-to-right), and \code{RL} (right-to-left).}
  \item{shape}{a character vector specifying the shape of nodes. The default value is "ellipse". See the polygon-based nodes in \code{graphviz} for more options.}
  \item{nodeOnly}{a logical vector specifying whether to display only the node number.}
}
}
\examples{
set.seed(1)
dat <- simu(100, 0, 1.3)
library(survival)
system.time(fit <- rocTree(Surv(Time, death) ~ z1 + z2, id = id,
data = dat, control = list(prune = TRUE, nflds = 10)))
plot(fit)
plot(fit, control = list(rankdir = "LR", nodeOnly = TRUE))
plot(fit, output = "visNetwork", control = list(nodeOnly = TRUE, shape = "box"))
}
\seealso{
See \code{\link{rocTree}} for creating \code{rocTree} objects.
}
