% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rocTree.R
\name{predict.rocTree}
\alias{predict.rocTree}
\title{Predicting based on a \code{rocTree} model.}
\usage{
\method{predict}{rocTree}(x, newdata, type = c("survival", "hazard"), ...)
}
\arguments{
\item{x}{an \code{rocTree} object.}

\item{newdata}{an optional data frame in which to look for variables with which to predict.
If omitted, the fitted predictors are used.}
}
\description{
The function gives predicted values.
}
