% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocForest.R
\name{rocForest}
\alias{rocForest}
\title{ROC-guided Regression Forest}
\usage{
rocForest(formula, data, id, subset, control = list())
}
\arguments{
\item{formula}{a formula object, with the response on the left of a '~' operator,
and the terms on the right. The response must be a survival object returned by the
'Surv' function.}

\item{data}{an optional data frame in which to interpret the variables occurring
in the 'formula'.}

\item{id}{an optional vector used to identify time dependent covariate.
If missing, then each individual row of 'data' is presumed to represent a distinct
subject and each covariate is treated as a baseline covariate. The length of 'id'
should be the same as the number of observations.}

\item{subset}{an optional vector specifying a subset of observations to be used in
the fitting process.}

\item{control}{a list of control parameters. See 'details' for important special
features of control parameters.}
}
\value{
An object of S3 class "\code{rocForest}" representing the fit, with the following components:
}
\description{
Fits a "\code{rocForest}" model.
}
\examples{
library(survival)
set.seed(123)
dat <- simu(40, 0, 1.1)
fit <- rocForest(Surv(Y, death) ~ z1 + z2, id = id, data = dat,
    control = list(minsp = 3, minsp2 = 1))
fit

## Print individual trees
print(fit, 1)
print(fit, 2)

}
