% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocTree.R
\name{rocTree}
\alias{rocTree}
\title{ROC-guided Regression Trees}
\usage{
rocTree(formula, data, id, subset, control = list())
}
\arguments{
\item{formula}{a formula object, with the response on the left of a '~' operator,
and the terms on the right. The response must be a survival object returned by the
'Surv' function.}

\item{data}{an optional data frame in which to interpret the variables occurring
in the 'formula'.}

\item{id}{an optional vector used to identify time dependent covariate.
If missing, then each individual row of 'data' is presumed to represent a distinct
subject and each covariate is treated as a baseline covariate. The length of 'id'
should be the same as the number of observations.}

\item{subset}{an optional vector specifying a subset of observations to be used in
the fitting process.}

\item{control}{a list of control parameters. See 'details' for important special
features of control parameters.}
}
\value{
An object of S3 class "\code{rocTree}" representing the fit, with the
following components:
\describe{
\item{Frame}{is a data frame describe the resulting tree.}
\item{r2Final}{estimated hazards at all terminal nodes.}
\item{tt}{time values to estimate hazard. Default length is 500. }
}
}
\description{
Fits a "\code{rocTree}" model.
}
\details{
The argument "control" defaults to a list with the following values:
\describe{
  \item{\code{tau}}{maximum follow-up time; default value is 0.4.}
  \item{\code{M}}{maximum node number allowed to be in the tree; the default value is 1000.}
  \item{\code{hN}}{smoothing parameter; the default value is "tau / 20".}
  \item{\code{minsp}, \code{minsp2}}{The interval (\code{minsp2}, \code{minsp}) denotes the
range for the number of the minimum risk observations required to split; the default value is (5, 20).}
  \item{\code{disc}}{a logical vector specifying whether the input covariate are discrete (\code{disc} = 0).
The length of "disc" should be the same as the number of covariates.}
  \item{\code{CV}}{a logical vector specifying whether a cross-validation is performed; the default value is FALSE.}
  \item{\code{nflds}}{number of folds; the default value is 10. This argument is only needed if \code{CV} = TRUE.}
  \item{\code{Trace}}{a logical vector specifying whether to display the splitting path; the default value is FALSE.}
  \item{\code{parallel}}{a logical vector specifying whether parallel computing is applied when \code{CV} = TRUE;
the default value is FALSE.}
  \item{\code{parCluster}}{an integer value specifying the number of CPU cores to be used when \code{CV} = TRUE and
\code{parallel} = TRUE. The default value is half of the CPU cores detected.}
}
}
\examples{
library(survival)
set.seed(123)
dat <- simu(40, 0, 1.1)
fit <- rocTree(Surv(Y, death) ~ z1 + z2, id = id, data = dat, control = list(CV = TRUE))
fit
}
\references{
Sun Y. and Wang, M.C. (2018+). ROC-guided classification and survival trees. \emph{Technical report}.
}
\seealso{
See \code{\link{print.rocTree}} and \code{\link{plot.rocTree}} for printing and plotting an \code{rocTree}, respectively.
}
\keyword{rocTree}
